third_party/blink/web_tests/editing/execCommand/remove-list-from-multi-list-items-expected.txt

This test verifies removing of an (Un)OrderedList on a selection consisting of multiple list items (paragaraphs) when the end point of the selection points to the end of the last list item; followed by another paragraph not part of the selection/list.
| "
"
| <p>
|   id="start"
|   "<#selection-anchor>This content should"
|   <br>
|   "not be displayed as"
|   <br>
|   "an unordered list.<#selection-focus>"
| "
"
| <p>
|   "This is some content that should not be added to the selection or the list."
| "
"
